Surface score components and malcs deetz

+++ internal/engine/eval/trusty/trusty.go
@@ -244,7 +244,7 @@ func classifyDependency(
 	// Ensure don't panic checking all fields are there
 	for _, fld := range []string{"activity", "provenance"} {
 		if _, ok := descr[fld]; !ok {
-			descr[fld] = 0
+			descr[fld] = float64(0)
 		}
 	}
